requirements.txt
django>=5.0
mysql-connector-python
django-crispy-forms
crispy-bootstrap5
django-cleanup